[COUR-26] Introducing controller interface for course as pilot and comment all the other controller
